Move QtTranslationSaxParser into QtTranslationParser as nested static class
